+++ hook-core/hooks/branch-protect.sh
@@ -1,0 +1,1 @@
+../../hooks/branch-protect.sh