+++ hook-core/scripts/devgate/require-rci-update-if-p0p1.sh
@@ -1,0 +1,1 @@
+../../../scripts/devgate/require-rci-update-if-p0p1.sh