+++ hook-core/scripts/devgate/view-snapshot.sh
@@ -1,0 +1,1 @@
+../../../scripts/devgate/view-snapshot.sh